% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidphinfodemic.R
\docType{package}
\name{covidphinfodemic}
\alias{covidphinfodemic}
\alias{covidphinfodemic-package}
\title{Utilities to Search and Extract Metadata from Philippines-Focused
Infographics and Dashboards}
\description{
The COVID-19 pandemic can be characterised by the epidemiological
phenomenon and transmission dynamics of this novel coronavirus causing not
only significant global health consequences but also societal and economic
impacts. However, a much less recognised yet equally significant phenomenon
of COVID-19 infodemic characterised as the excessive amount of information
about the pandemic, which makes the identification of a solution difficult
and can spread misinformation, disinformation and rumours during the health
emergency has been co-existing. The infodemic has hampered an effective
public health response and hasa created confusion and distrust among people.
We aim to document and analyse this phenomenon in the context of the
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

Authors:
\itemize{
  \item Chris Mercado \email{cegmercado@gmail.com} (\href{https://orcid.org/0000-0002-5588-9543}{ORCID})
  \item Robert Medina \email{jcmedina1@up.edu.ph} (\href{https://orcid.org/0000-0001-6647-5453}{ORCID})
}

}
\keyword{internal}
