% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dashboards}
\alias{dashboards}
\title{Dashboards on COVID-19 specific to the Philippines}
\format{
A tibble with 152 rows and 6 columns
}
\usage{
dashboards
}
\description{
Dashboards on COVID-19 specific to the Philippines
}
\examples{
dashboards

}
\keyword{dataset}
