% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{get_db_text}
\alias{get_db_text}
\title{Extract text from actual dashboard websites}
\usage{
get_db_text(url)
}
\arguments{
\item{url}{A vector of dashboard websites}
}
\value{
A vector of text contents of dashboard websites
}
\description{
Extract text from actual dashboard websites
}
\examples{
get_db_text(url = "https://covid19.mathdro.id")

}
