% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{get_gh_readme}
\alias{get_gh_readme}
\title{Extract text from README of GitHub repositories}
\usage{
get_gh_readme(repo)
}
\arguments{
\item{repo}{A vector of GitHub repository URLs}
}
\value{
A vector of text contents of README files
}
\description{
Extract text from README of GitHub repositories
}
\examples{
get_gh_readme(repo = "https://github.com/como-ph/covidphinfodemic")

}
