% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{githubSearch}
\alias{githubSearch}
\title{GitHub search results for COVID-19 dashboards regarding the Philippines}
\format{
A tibble with 126 rows and 97 columns
}
\usage{
githubSearch
}
\description{
This dataset is produced using the \link{search_gh_covidph} function.
}
\examples{
githubSearch

}
\keyword{dataset}
