% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_gh_covidph}
\alias{search_gh_covidph}
\title{Perform a search for Philippines- and COVID-19-related GitHub repositories}
\usage{
search_gh_covidph(query = NULL, per_page = 100, page = NULL)
}
\arguments{
\item{query}{GitHub query.}

\item{per_page}{How many results per page should be returned. Defaults to
100 (maximum)}

\item{page}{Results page to return}
}
\value{
A tibble of search results
}
\description{
Perform a search for Philippines- and COVID-19-related GitHub repositories
}
\examples{
\dontrun{
  search_gh_covidph(query = "covid19+philippines+in:readme in:description in:name",
                    page = 1)
}

}
