% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{set_gh_token}
\alias{set_gh_token}
\title{Set GitHub authentication credentials}
\usage{
set_gh_token(token)
}
\arguments{
\item{token}{GitHub personal access token (PAT)}
}
\value{
GitHub system environment variables set.
}
\description{
Set GitHub authentication credentials
}
\examples{
\dontrun{
  set_gh_token(token = "XXXXXXXXXXXXXXXX")
}

}
